# packages/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentDetail": {
        "description": "Detail page payload \u2014 same metrics + recent traces + violation\nbreakdown by policy.",
        "properties": {
          "active_traces": {
            "default": 0,
            "title": "Active Traces",
            "type": "integer"
          },
          "activity": {
            "title": "Activity",
            "type": "string"
          },
          "activity_buckets": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Activity Buckets",
            "type": "array"
          },
          "avg_duration_ms": {
            "title": "Avg Duration Ms",
            "type": "integer"
          },
          "error_rate": {
            "title": "Error Rate",
            "type": "number"
          },
          "has_violations": {
            "title": "Has Violations",
            "type": "boolean"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "providers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "recent_traces": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Trace"
            },
            "title": "Recent Traces",
            "type": "array"
          },
          "seconds_since_last_seen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds Since Last Seen"
          },
          "top_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Model"
          },
          "top_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Provider"
          },
          "total_cost_usd": {
            "title": "Total Cost Usd",
            "type": "number"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          },
          "trace_count": {
            "title": "Trace Count",
            "type": "integer"
          },
          "violation_count": {
            "title": "Violation Count",
            "type": "integer"
          },
          "violations_by_policy": {
            "additionalProperties": true,
            "default": {},
            "title": "Violations By Policy",
            "type": "object"
          },
          "violations_by_severity": {
            "additionalProperties": true,
            "default": {},
            "title": "Violations By Severity",
            "type": "object"
          }
        },
        "required": [
          "name",
          "project",
          "trace_count",
          "total_cost_usd",
          "total_tokens",
          "avg_duration_ms",
          "error_rate",
          "violation_count",
          "has_violations",
          "last_seen",
          "top_model",
          "top_provider",
          "seconds_since_last_seen",
          "activity"
        ],
        "title": "AgentDetail",
        "type": "object"
      },
      "AgentListResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentSummary"
            },
            "title": "Agents",
            "type": "array"
          },
          "older_data_exists": {
            "default": false,
            "title": "Older Data Exists",
            "type": "boolean"
          },
          "projects": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Projects",
            "type": "array"
          },
          "window_hours": {
            "title": "Window Hours",
            "type": "integer"
          }
        },
        "required": [
          "agents",
          "window_hours"
        ],
        "title": "AgentListResponse",
        "type": "object"
      },
      "AgentSummary": {
        "description": "One card on the agent grid.",
        "properties": {
          "active_traces": {
            "default": 0,
            "title": "Active Traces",
            "type": "integer"
          },
          "activity": {
            "title": "Activity",
            "type": "string"
          },
          "activity_buckets": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Activity Buckets",
            "type": "array"
          },
          "avg_duration_ms": {
            "title": "Avg Duration Ms",
            "type": "integer"
          },
          "error_rate": {
            "title": "Error Rate",
            "type": "number"
          },
          "has_violations": {
            "title": "Has Violations",
            "type": "boolean"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "providers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "seconds_since_last_seen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds Since Last Seen"
          },
          "top_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Model"
          },
          "top_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Provider"
          },
          "total_cost_usd": {
            "title": "Total Cost Usd",
            "type": "number"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          },
          "trace_count": {
            "title": "Trace Count",
            "type": "integer"
          },
          "violation_count": {
            "title": "Violation Count",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "project",
          "trace_count",
          "total_cost_usd",
          "total_tokens",
          "avg_duration_ms",
          "error_rate",
          "violation_count",
          "has_violations",
          "last_seen",
          "top_model",
          "top_provider",
          "seconds_since_last_seen",
          "activity"
        ],
        "title": "AgentSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first AI agent observability \u2014 ingest and query API",
    "title": "Korveo API",
    "version": "0.6.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agents": {
      "get": {
        "description": "List all distinct agents (= distinct trace.name) and their\nactivity metrics over the configured window.\n\nThe query is one DuckDB aggregate \u2014 no per-agent N+1. Joins against\nspans for cost / token / model / error_rate, against\npolicy_violations for the violation count.",
        "operationId": "list_agents_v1_agents_get",
        "parameters": [
          {
            "description": "Activity window for aggregated metrics",
            "in": "query",
            "name": "window_hours",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Activity window for aggregated metrics",
              "maximum": 720,
              "minimum": 1,
              "title": "Window Hours",
              "type": "integer"
            }
          },
          {
            "description": "Substring filter on agent name",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring filter on agent name",
              "title": "Search"
            }
          },
          {
            "description": "Filter by framework: openclaw / mastra / voltagent / default",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by framework: openclaw / mastra / voltagent / default",
              "title": "Project"
            }
          },
          {
            "description": "Filter by LLM provider: anthropic / openai / ollama / etc.",
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by LLM provider: anthropic / openai / ollama / etc.",
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agents"
      }
    },
    "/v1/agents/{agent_name}": {
      "get": {
        "description": "One agent's detail page: same metrics as the card, plus the\nmost-recent traces and a per-policy violation breakdown so the\noperator can scan what's been firing.",
        "operationId": "get_agent_v1_agents__agent_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "title": "Agent Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window_hours",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 720,
              "minimum": 1,
              "title": "Window Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent"
      }
    },
...
  }
}